Rename fold name
