+++ spine-cocos-creator/AttachmentVertices.cpp
