+++ spine-cocos-creator/AttachmentVertices.h
