+++ spine-cocos-creator/CreatorAttachmentLoader.cpp
