+++ spine-cocos-creator/CreatorAttachmentLoader.h
