+++ spine-cocos-creator/SpineAnimation.cpp
