+++ spine-cocos-creator/SpineAnimation.h
