+++ spine-cocos-creator/SpineRenderer.cpp
