+++ spine-cocos-creator/SpineRenderer.h
