+++ spine-cocos-creator/spine-cocos2dx.cpp
