--- spine-cocoscreator/SkeletonDataMgr.h
@@ -1,68 +1,0 @@
-/****************************************************************************
- Copyright (c) 2019 Xiamen Yaji Software Co., Ltd.
- 
- http://www.cocos2d-x.org
- 
- Permission is hereby granted, free of charge, to any person obtaining a copy
- of this software and associated documentation files (the "Software"), to deal
- in the Software without restriction, including without limitation the rights
- to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- copies of the Software, and to permit persons to whom the Software is
- furnished to do so, subject to the following conditions:
- 
- The above copyright notice and this permission notice shall be included in
- all copies or substantial portions of the Software.
- 
- THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- THE SOFTWARE.
- ****************************************************************************/
-
-#pragma once
-
-#include <map>
-#include <string>
-#include "spine/SkeletonData.h"
-#include "base/CCRef.h"
-#include "spine/spine.h"
-
-namespace spine {
-
-/**
- * Cache skeleton data.
- */
-class SkeletonDataMgr {
-public:
-    static SkeletonDataMgr* getInstance ()
-    {
-        if (_instance == nullptr)
-        {
-            _instance = new SkeletonDataMgr();
-        }
-        return _instance;
-    }
-    
-    static void destroyInstance ()
-    {
-        if (_instance)
-        {
-            delete _instance;
-            _instance = nullptr;
-        }
-    }
-    
-    SkeletonDataMgr () {}
-    virtual ~SkeletonDataMgr () {}
-    bool hasSkeletonData (const std::string& uuid);
-    void setSkeletonData (const std::string& uuid, spSkeletonData* data, spAtlas* atlas, spAttachmentLoader* attachmentLoader);
-    spSkeletonData* retainByUUID (const std::string& uuid);
-    void releaseByUUID (const std::string& uuid);
-private:
-    static SkeletonDataMgr* _instance;
-};
-
-}
